requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==45.0.4
Django==5.2.3
djangorestframework==3.16.0
exceptiongroup==1.3.0
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
pycparser==2.22
Pygments==2.19.1
PyMySQL==1.1.1
pytest==8.4.0
pytest-django==4.11.1
python-dotenv==1.1.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.0
